@@ -5,14 +5,14 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.2",
-        "ext-ldap": "*",
-        "directorytree/ldaprecord-laravel": "^v2.1.1",
-        "laravel/lumen-framework": "^5.6|^6.0|^7.0|^8.0|^9.0"
+        "php": ">=8.1",
+        "directorytree/ldaprecord-laravel": "^v3.0.x-dev",
+        "laravel/lumen-framework": "^8.0|^9.0|^v10.0"
     },
     "require-dev": {
         "mockery/mockery": "^1.3.1",
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^8.5",
+        "laravel/pint": "^1.10"
     },
     "archive": {
         "exclude": ["/tests"]
